move check-versions under wasp command
