+++ tools/wasp-cli/waspcmd/cmd.go
@@ -26,6 +26,8 @@ func Init(rootCmd *cobra.Command) {
 	rootCmd.AddCommand(waspNodesCmd)
 
 	waspNodesCmd.AddCommand(initAddWaspNodeCmd())
+	waspNodesCmd.AddCommand(initCheckVersionsCmd())
+
 }
 
 func initAddWaspNodeCmd() *cobra.Command {
